replace uglify plugin with terser plugin
