+++ package.json
@@ -137,9 +137,10 @@
     "stylelint": "^13.13.1",
     "stylelint-checkstyle-formatter": "^0.1.2",
     "stylelint-scss": "^3.20.1",
+    "terser-webpack-plugin": "^5.2.4",
     "testcafe": "^1.14.2",
     "time-grunt": "^2.0.0",
-    "uglifyjs-webpack-plugin": "^2.2.0",
+    "uglify-js": "^3.14.2",
     "wait-on": "^3.3.0",
     "webdriver-manager": "^13.0.0",
     "webpack": "^5.51.1",
