+++ scripts/buildtools/webpack/plugins.js
@@ -1,10 +1,7 @@
-/*eslint camelcase: ["error", {properties: "never"}]*/
-
 const { readFileSync } = require('fs');
 const { join } = require('path');
 const { DefinePlugin, BannerPlugin, IgnorePlugin } = require('webpack');
 const { BundleAnalyzerPlugin } = require('webpack-bundle-analyzer');
-const UglifyJsPlugin = require('uglifyjs-webpack-plugin');
 
 function webpackBundleAnalyzer(reportFilename = 'okta-sign-in.analyzer') {
   return new BundleAnalyzerPlugin({
@@ -26,57 +23,6 @@ function prodMode() {
 function devMode() {
   return new DefinePlugin({
     DEBUG: true
-  });
-}
-
-function uglify() {
-  return new UglifyJsPlugin({
-    uglifyOptions: {
-      compress: {
-        // Drop all console.* and Logger statements
-        drop_console: true,
-        drop_debugger: true,
-        pure_funcs: [
-          'Logger.trace',
-          'Logger.dir',
-          'Logger.time',
-          'Logger.timeEnd',
-          'Logger.group',
-          'Logger.groupEnd',
-          'Logger.assert',
-          'Logger.log',
-          'Logger.info',
-          'Logger.warn',
-          'Logger.deprecate'
-        ],
-      },
-      sourceMap: true,
-      comments: (node, comment) => {
-        // Remove other Okta copyrights
-        const isLicense = /^!/.test(comment.value) ||
-                        /.*(([Ll]icense)|([Cc]opyright)|(\([Cc]\))).*/.test(comment.value);
-        const isOkta = /.*Okta.*/.test(comment.value);
-
-        // Some licenses are in inline comments, rather than standard block comments.
-        // UglifyJS2 treats consecutive inline comments as separate comments, so we
-        // need exceptions to include all relevant licenses.
-        const exceptions = [
-          'Chosen, a Select Box Enhancer',
-          'by Patrick Filler for Harvest',
-          'Version 0.11.1',
-          'Full source at https://github.com/harvesthq/chosen',
-
-          'Underscore.js 1.8.3'
-        ];
-
-        const isException = exceptions.some(exception => {
-          return comment.value.indexOf(exception) !== -1;
-        });
-
-        return (isLicense || isException) && !isOkta;
-      },
-      warnings: false
-    }
   });
 }
 
@@ -109,12 +55,11 @@ function failOnBuildFail() {
 function plugins(options = {}) {
   const { isProduction, skipAnalyzer } = options;
   const list = isProduction ? 
-    // Uglify and add license header
+    // Add license header
     [
       failOnBuildFail(),
       emptyModule(),
       prodMode(),
-      uglify(),
       banner(),
     ] : 
     // Use DEBUG/development environment w/ console warnings
