+++ webpack.common.config.js
@@ -1,4 +1,7 @@
+/* eslint-disable camelcase */
 var path      = require('path');
+const TerserPlugin = require('terser-webpack-plugin');
+
 var SRC = path.resolve(__dirname, 'src');
 var TARGET_JS = path.resolve(__dirname, 'target/js/');
 var LOCAL_PACKAGES = path.resolve(__dirname, 'packages/');
@@ -78,6 +81,73 @@ module.exports = function(outputFilename, mode = 'development') {
     // Webpack attempts to add a polyfill for process
     // and setImmediate, because q uses process to see
     // if it's in a Node.js environment
-    node: false
+    node: false,
+
+    optimization: {
+      minimize: true,
+      minimizer: [
+        new TerserPlugin({
+          minify: (file, sourceMap) => {
+            // https://github.com/mishoo/UglifyJS2#minify-options
+            const uglifyJsOptions = {
+              compress: {
+                // Drop all console.* and Logger statements
+                drop_console: true,
+                drop_debugger: true,
+                pure_funcs: [
+                  'Logger.trace',
+                  'Logger.dir',
+                  'Logger.time',
+                  'Logger.timeEnd',
+                  'Logger.group',
+                  'Logger.groupEnd',
+                  'Logger.assert',
+                  'Logger.log',
+                  'Logger.info',
+                  'Logger.warn',
+                  'Logger.deprecate'
+                ],
+              },
+              sourceMap: true,
+              output: {
+                comments: (node, comment) => {
+                  // Remove other Okta copyrights
+                  const isLicense = /^!/.test(comment.value) ||
+                                  /.*(([Ll]icense)|([Cc]opyright)|(\([Cc]\))).*/.test(comment.value);
+                  const isOkta = /.*Okta.*/.test(comment.value);
+          
+                  // Some licenses are in inline comments, rather than standard block comments.
+                  // UglifyJS2 treats consecutive inline comments as separate comments, so we
+                  // need exceptions to include all relevant licenses.
+                  const exceptions = [
+                    'Chosen, a Select Box Enhancer',
+                    'by Patrick Filler for Harvest',
+                    'Version 0.11.1',
+                    'Full source at https://github.com/harvesthq/chosen',
+          
+                    'Underscore.js 1.8.3'
+                  ];
+          
+                  const isException = exceptions.some(exception => {
+                    return comment.value.indexOf(exception) !== -1;
+                  });
+          
+                  return (isLicense || isException) && !isOkta;
+                },
+              },
+              warnings: false
+            };
+  
+            if (sourceMap) {
+              uglifyJsOptions.sourceMap = {
+                content: sourceMap,
+              };
+            }
+  
+            return require('uglify-js').minify(file, uglifyJsOptions);
+          },
+        }),
+      ],
+    },
   };
 };
